refactor(workspace): extract explorer plugin parts
